+++ source/Nuke.GlobalTool/ProjectUpdater.cs
@@ -14,6 +14,8 @@ namespace Nuke.GlobalTool;
 
 public static class ProjectUpdater
 {
+    public const string TARGET_FRAMEWORK = "net10.0";
+
     public static void Update(string projectFile)
     {
         var buildProject = ProjectModelTasks.ParseProject(projectFile).NotNull();
@@ -29,15 +31,15 @@ public static class ProjectUpdater
 
     private static void UpdateTargetFramework(Microsoft.Build.Evaluation.Project buildProject)
     {
-        buildProject.SetProperty("TargetFramework", "net10.0");
+        buildProject.SetProperty("TargetFramework", TARGET_FRAMEWORK);
     }
 
     private static void UpdateNukeCommonPackage(Microsoft.Build.Evaluation.Project buildProject, out FloatRange previousPackageVersion)
     {
-        var packageItem = buildProject.Items.SingleOrDefault(x => x.EvaluatedInclude == Constants.NukeCommonPackageId).NotNull();
+        var packageItem = buildProject.Items.SingleOrDefault(x => x.EvaluatedInclude == Constants.NukePackageId).NotNull();
         previousPackageVersion = FloatRange.Parse(packageItem.GetMetadataValue("Version"));
 
-        var latestPackageVersion = NuGetVersionResolver.GetLatestVersion(Constants.NukeCommonPackageId, includePrereleases: false).GetAwaiter().GetResult();
+        var latestPackageVersion = NuGetVersionResolver.GetLatestVersion(Constants.NukePackageId, includePrereleases: false).GetAwaiter().GetResult();
         if (previousPackageVersion.Satisfies(NuGetVersion.Parse(latestPackageVersion)))
             return;